+++ deeprl/environments/__init__.py
@@ -1,2 +1,7 @@
-from .gymnasium_env_wrapper import GymnasiumEnvWrapper
 from .base_environment import BaseEnvironment
+from .gymnasium_env_wrapper import GymnasiumEnvWrapper
+
+__all__ = [
+    'BaseEnvironment', 
+    'GymnasiumEnvWrapper'
+]