+++ deeprl/policies/__init__.py
@@ -1,4 +1,12 @@
+# policies/__init__.py
 from .base_policy import BasePolicy
-from .softmax_policy import SoftmaxPolicy
 from .deterministic_policy import DeterministicPolicy
 from .epsilon_greedy_policy import EpsilonGreedyPolicy
+from .softmax_policy import SoftmaxPolicy
+
+__all__ = [
+    'BasePolicy', 
+    'DeterministicPolicy', 
+    'EpsilonGreedyPolicy', 
+    'SoftmaxPolicy'
+]